+++ crates/api-desc/src/platform.rs
@@ -31,16 +31,32 @@ pub(crate) fn new() -> Item {
         update::new(),
         #[cfg(feature = "api-platform")]
         item! {
+            /// Returns the serial of the platform.
+            ///
+            /// Returns the length of the serial in bytes. The serial is not allocated if the
+            /// length is zero (and the pointer is not written).
+            fn serial "ps" {
+                /// Where to write the serial.
+                ///
+                /// If the returned length is positive, the (inner) pointer will be allocated by the
+                /// callee and must be freed by the caller. It is thus owned by the caller when the
+                /// function returns.
+                ptr: *mut *mut u8,
+            } -> usize
+        },
+        #[cfg(feature = "api-platform")]
+        item! {
             /// Returns the version of the platform.
             ///
-            /// Returns the length of the version in bytes. This may be larger than the capacity, in
-            /// which case only a prefix was written.
+            /// Returns the length of the version in bytes. The version is not allocated if the
+            /// length is zero (and the pointer is not written).
             fn version "pv" {
                 /// Where to write the version.
-                ptr: *mut u8,
-
-                /// Capacity of the buffer.
-                len: usize,
+                ///
+                /// If the returned length is positive, the (inner) pointer will be allocated by the
+                /// callee and must be freed by the caller. It is thus owned by the caller when the
+                /// function returns.
+                ptr: *mut *mut u8,
             } -> usize
         },
         #[cfg(feature = "api-platform")]