+++ crates/interpreter/test.sh
@@ -22,10 +22,11 @@ ensure_submodule third_party/WebAssembly/threads
 
 test_helper
 
-cargo check
-cargo check --target=thumbv7em-none-eabi
-cargo check --target=riscv32imc-unknown-none-elf \
+cargo test --lib --features=toctou
+cargo check --lib --target=thumbv7em-none-eabi
+cargo check --lib --target=riscv32imc-unknown-none-elf \
   --features=portable-atomic/critical-section
 RUSTFLAGS=--cfg=portable_atomic_unsafe_assume_single_core \
-  cargo check --target=riscv32imc-unknown-none-elf
-cargo test --features=debug,toctou,float-types,vector-types,threads
+  cargo check --lib --target=riscv32imc-unknown-none-elf
+cargo test --test=spec --features=debug,toctou,float-types,vector-types,threads
+cargo check --example=hello