+++ crates/protocol-usb/test.sh
@@ -19,7 +19,7 @@ set -e
 
 test_helper
 
-cargo test --features=host
-cargo check --target=thumbv7em-none-eabi --features=device
-cargo check --features=host,std,log
-DEFMT_LOG=trace cargo check --target=thumbv7em-none-eabi --features=device,defmt
+cargo test --lib --features=host
+cargo check --lib --target=thumbv7em-none-eabi --features=device
+cargo check --lib --features=host,std,log
+DEFMT_LOG=trace cargo check --lib --target=thumbv7em-none-eabi --features=device,defmt