+++ crates/runner-nordic/test.sh
@@ -21,8 +21,8 @@ ensure_applet
 
 test_helper
 
-cargo check --target=thumbv7em-none-eabi --features=wasm,debug
-cargo check --target=thumbv7em-none-eabi --features=wasm,debug,software-crypto
-DEFMT_LOG=trace cargo check --target=thumbv7em-none-eabi --features=wasm,debug
-cargo check --target=thumbv7em-none-eabi --features=wasm,release
-cargo check --target=thumbv7em-none-eabi --features=native,release
+cargo check --bin=runner-nordic --target=thumbv7em-none-eabi --features=wasm,debug
+cargo check --bin=runner-nordic --target=thumbv7em-none-eabi --features=wasm,debug,software-crypto
+DEFMT_LOG=trace cargo check --bin=runner-nordic --target=thumbv7em-none-eabi --features=wasm,debug
+cargo check --bin=runner-nordic --target=thumbv7em-none-eabi --features=wasm,release
+cargo check --bin=runner-nordic --target=thumbv7em-none-eabi --features=native,release