+++ scripts/ci-changelog.sh
@@ -19,6 +19,8 @@ set -e
 
 # This script checks that Cargo.toml and CHANGELOG.md files are correct.
 
+INCLUDE='["/LICENSE", "/src/"]'
+LICENSE="$(readlink -f LICENSE)"
 for dir in $(find crates -name Cargo.toml -printf '%h\n' | sort); do
   ( cd $dir
     publish="$(package_publish)"
@@ -26,7 +28,8 @@ for dir in $(find crates -name Cargo.toml -printf '%h\n' | sort); do
     [ -e test.sh ] || e "test.sh for $dir is missing"
     $publish || exit 0
     [ -e CHANGELOG.md ] || e "CHANGELOG.md for $dir is missing"
-    [ "$(package_include)" = '["/src"]' ] || e "Cargo.toml should only include the src directory"
+    [ "$(package_include)" = "$INCLUDE" ] || e "Cargo.toml should include exactly $INCLUDE"
+    [ "$(readlink -f LICENSE)" = "$LICENSE" ] || e "LICENSE is not a symlink to the top-level one"
     [ -z "$(package_exclude)" ] || e "Cargo.toml should not exclude anything"
     ref=$(git log -n1 --pretty=format:%H origin/main.. -- CHANGELOG.md)
     [ -n "$ref" ] || ref=origin/main