+++ scripts/publish.sh
@@ -23,11 +23,11 @@ set -e
 
 TOPOLOGICAL_ORDER=(
   logger
+  wire-derive
   error
+  wire
   sync
   protocol
-  cli-tools
-  cli
   interpreter
   store
   api-desc
@@ -38,6 +38,8 @@ TOPOLOGICAL_ORDER=(
   board
   scheduler
   protocol-usb
+  cli-tools
+  cli
 )
 
 listed_crates() {
@@ -92,6 +94,6 @@ for crate in "${TOPOLOGICAL_ORDER[@]}"; do
       exit
     fi
     i "Publish $crate from $latest to $current"
-    eval "$(sed -n 's/^cargo \(check\|test\)/cargo publish/p;T;q' test.sh)"
+    eval "$(sed -En 's/^cargo (check|test) --(lib|bin=[^ ]*)/cargo publish/p;T;q' test.sh)"
   )
 done