+++ scripts/sync.sh
@@ -43,12 +43,16 @@ for dir in $(find crates -name Cargo.toml -printf '%h\n' | sort); do
   esac
   add_lint $file warn rust.unsafe-op-in-unsafe-fn
   case $crate in
-    # TODO: Remove api-desc once the main.rs logic is moved to xtask.
-    api-desc|*/fuzz) ;;
+    */fuzz) ;;
     *) add_lint $file warn rust.unused-crate-dependencies ;;
   esac
   # add_lint $file warn rust.unused-results
 done
+
+( cd crates/protocol/crates/schema
+  cargo run --features=host
+  cargo run --features=device
+)
 
 book_example() {
   local src=book/src/applet/prelude/$1.rs