+++ scripts/wrapper.sh
@@ -54,7 +54,7 @@ case "$1" in
     esac
     ;;
   mdbook) ensure_cargo mdbook 0.4.40 ;;
-  probe-rs) ensure_cargo probe-rs 0.24.0 --features=cli ;;
+  probe-rs) ensure_cargo probe-rs-tools 0.24.0 ;;
   rust-objcopy|rust-size) ensure_cargo cargo-binutils 0.3.6 ;;
   taplo) ensure_cargo taplo-cli 0.9.0 ;;
   twiggy) ensure_cargo twiggy 0.7.0 ;;