+++ crates/logger/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Patch
 
+- Publish LICENSE file
+- Depend on `alloc` to silence `unused-crate-dependencies` false positive
 - Update dependencies
 - Use common Wasefire lints