+++ crates/sync/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## 0.1.1-git
 
+### Minor
+
+- Add `Lazy` as a re-export of `spin::Lazy`
+
 ### Patch
 
+- Publish LICENSE file
+- Depend on `alloc` to silence `unused-crate-dependencies` false positive
 - Use common Wasefire lints
 
 ## 0.1.0